requirements.txt
requests==2.32.3
python-dotenv==1.1.0
python-telegram-bot==13.0
urllib3==1.26.15; python_version >= '3.11'
standard-imghdr; python_version >= '3.13'
